Bug fixes
Fixed a bug where sometimes the players' values would be in the millions. This was due to accessing an empty array cell. To ensure this won't happen again, the cards data structure in Hand was changed to a std::vector and all fo the for loops were changed to for each.